revise to use decorator indicate lyric source
